consolidates mock data
