--- libs/core/ui/.storybook/mock-data/checkbox-states.ts
@@ -1,22 +1,0 @@
-import { CheckboxStates } from '../../src/lib/nested-checkboxes/nested-checkboxes.component';
-
-export const SOME_SELECTED: CheckboxStates = {
-  Africa: 'indeterminate',
-  Morocco: 'indeterminate',
-  'Southern Africa': 'indeterminate',
-  Swaziland: 'checked',
-  'Northern Africa': 'indeterminate',
-  Fes: 'checked',
-};
-
-export const ALL_SELECTED: CheckboxStates = {
-  Africa: 'checked',
-  'Southern Africa': 'checked',
-  Swaziland: 'checked',
-  Namibia: 'checked',
-  'Central Africa': 'checked',
-  'Northern Africa': 'checked',
-  Morocco: 'checked',
-  Marrakesh: 'checked',
-  Fes: 'checked',
-};
