--- libs/core/ui/.storybook/mock-data/nested-item.ts
@@ -1,31 +1,0 @@
-
-export interface NestedItem {
-  id: string;
-  children?: NestedItem[];
-}
-
-export const NESTED_ITEM: NestedItem = {
-  id: 'Africa',
-  children: [
-    {
-      id: 'Southern Africa',
-      children: [
-        { id: 'Swaziland' },
-        { id: 'Namibia' },
-      ],
-    },
-    { id: 'Central Africa' },
-    {
-      id: 'Northern Africa',
-      children: [
-        {
-          id: 'Morocco',
-          children: [
-            { id: 'Marrakesh' },
-            { id: 'Fes' },
-          ],
-        },
-      ],
-    },
-  ],
-};
